Moved header utility class
- We have relocated the Header utility class into the response directory, makes more sense with the Headers helper
