+++ app/Http/Controllers/CategoryController.php
@@ -7,7 +7,7 @@ use App\Option\Delete;
 use App\Option\Get;
 use App\Option\Patch;
 use App\Option\Post;
-use App\Utilities\Header;
+use App\Response\Header\Header;
 use App\Utilities\Pagination as UtilityPagination;
 use App\Utilities\RoutePermission;
 use App\Validators\Parameters;
