+++ app/Http/Controllers/IndexController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Utilities\Header;
+use App\Response\Header\Header;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Route;
