+++ app/Http/Controllers/ResourceTypeController.php
@@ -12,7 +12,7 @@ use App\Option\Get;
 use App\Option\Patch;
 use App\Option\Post;
 use App\Response\Cache;
-use App\Utilities\Header;
+use App\Response\Header\Header;
 use App\Utilities\Pagination as UtilityPagination;
 use App\Utilities\Request as UtilityRequest;
 use App\Utilities\RoutePermission;
