+++ app/Http/Controllers/Summary/RequestController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Summary;
 
 use App\Http\Controllers\Controller;
 use App\Option\Get;
-use App\Utilities\Header;
+use App\Response\Header\Header;
 use App\Validators\Parameters;
 use App\Models\Summary\RequestLog;
 use Illuminate\Http\JsonResponse;
