+++ app/Http/Controllers/Summary/ResourceController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Summary;
 
 use App\Http\Controllers\Controller;
 use App\Option\Get;
-use App\Utilities\Header;
+use App\Response\Header\Header;
 use App\Utilities\RoutePermission;
 use App\Validators\Route;
 use App\Models\Summary\Resource;
