+++ app/Response/Header/Header.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Utilities;
+namespace App\Response\Header;
 
 /**
  * Generate the headers for the request.
